# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Physical AI 3D API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/models/{version_id}/engineering": {
      "post": {
        "tags": [
          "engineering"
        ],
        "summary": "Ask The Engineer",
        "operationId": "ask_the_engineer_api_v1_models__version_id__engineering_post",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Retry-safe key",
              "title": "Idempotency-Key"
            },
            "description": "Retry-safe key"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdviceBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "List Engineering Reports",
        "operationId": "list_engineering_reports_api_v1_models__version_id__engineering_get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngineeringReportOut"
                  },
                  "title": "Response List Engineering Reports Api V1 Models  Version Id  Engineering Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdviceBody": {
        "properties": {
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "material_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "printer_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printer Profile Id"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionSelection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "AdviceBody",
        "description": "A question about the part, or nothing — then the engineer reviews the whole part."
      },
...
      "BoxRegion": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "box",
            "title": "Kind",
            "default": "box"
          },
          "min_mm": {
            "prefixItems": [
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              },
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              },
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              }
            ],
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Min Mm"
          },
          "max_mm": {
            "prefixItems": [
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              },
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              },
              {
                "type": "number",
                "maximum": 100000.0,
                "minimum": -100000.0
              }
            ],
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Max Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "min_mm",
          "max_mm"
        ],
        "title": "BoxRegion",
        "description": "An axis-aligned volume in world millimetres — what a rectangular drag becomes."
      },
...
      "EngineeringReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Version Id"
          },
          "material_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "region": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_version_id",
          "material_id",
          "question",
          "region",
          "report",
          "created_at"
        ],
        "title": "EngineeringReportOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "type"
        ],
        "title": "JobAccepted",
        "description": "202 response for every long-running operation (docs/03 §1)."
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "waiting_input",
          "succeeded",
          "failed",
          "canceled"
        ],
        "title": "JobStatus"
      },
      "LassoRegion": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "lasso",
            "title": "Kind",
            "default": "lasso"
          },
          "axis": {
            "type": "string",
            "enum": [
              "x",
              "y",
              "z"
            ],
            "title": "Axis"
          },
          "offset_mm": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": -100000.0,
            "title": "Offset Mm"
          },
          "depth_mm": {
            "type": "number",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Depth Mm",
            "default": 10.0
          },
          "points_mm": {
            "items": {
              "prefixItems": [
                {
                  "type": "number",
                  "maximum": 100000.0,
                  "minimum": -100000.0
                },
                {
                  "type": "number",
                  "maximum": 100000.0,
                  "minimum": -100000.0
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "maxItems": 256,
            "minItems": 3,
            "title": "Points Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "axis",
          "offset_mm",
          "points_mm"
        ],
        "title": "LassoRegion",
        "description": "A freehand outline, projected onto one plane of the model.\n\nThe polygon is 2-D in the plane's own axes (the two that are not `axis`), at the\nplane's offset along `axis`, extruded through `depth_mm`. That is exactly what a finger\nor a pencil draws on a face, and it is enough for the kernel to work with."
      },
...
      "RegionSelection": {
        "properties": {
          "region": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BoxRegion"
              },
              {
                "$ref": "#/components/schemas/LassoRegion"
              }
            ],
            "title": "Region",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "box": "#/components/schemas/BoxRegion",
                "lasso": "#/components/schemas/LassoRegion"
              }
            }
          },
          "target": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "surface_axis": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "x",
                  "y",
                  "z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Axis"
          },
          "surface_sign": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "+",
                  "-"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Sign"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "region"
        ],
        "title": "RegionSelection",
        "description": "What the client sends with a prompt: where, and on what."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}